requirements.txt
asgiref==3.8.1
Django==5.1
django-allauth==64.1.0
django-filter==24.3
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
